refactored and fixed linting errors
