+++ src/utils/constants.ts
@@ -166,7 +166,7 @@ export enum APP_NAMES {
   RECORDEDIT = 'recordedit',
   RECORDSET = 'recordset',
   VIEWER = 'viewer'
-};
+}
 
 export const CLASS_NAMES = {
   CONTENT_LOADED: '-chaise-post-load',
