+++ src/utils/faceting-utils.ts
@@ -45,3 +45,31 @@ export function getInitialFacetPanelOpen(config: RecordsetConfig, initialReferen
   }
   return res;
 }
+
+/**
+ * Change render order for facets based on 'newOrder'
+ * @param defaultFacetColumns default order ex. 'reference.facetColumns'
+ * @param newOrder order in which the facets should be rendered 
+ * @returns facet columns in newOrder
+ */
+export const reorderFacets = (defaultFacetColumns:any[], newOrder:string[]) => {
+  // Create a map to store the indices of elements in array `newOrder`
+  const indexMap = new Map();
+  for (let i = 0; i < newOrder.length; i++) {
+      indexMap.set(newOrder[i], i);
+  }
+
+  const facetsInNewOrder = defaultFacetColumns.map((item: any, index: number) => [item, index])
+
+  // Sort array `defaultFacetColumns` based on the indices in array `newOrder`
+  facetsInNewOrder.sort((a, b) => {
+      const indexA = indexMap.get(a[0].sourceObjectWrapper.name);
+      const indexB = indexMap.get(b[0].sourceObjectWrapper.name);
+      // If either element is not found in `newOrder`, move it to the end
+      if (indexA === undefined) return 1;
+      if (indexB === undefined) return -1;
+      return indexA - indexB;
+  });
+
+  return facetsInNewOrder;
+}
